PANIC HAS DESCENDED UPON US

+++ Ventura13/Assets/Script/ShipHealth.cs
@@ -150,7 +150,7 @@ public class ShipHealth : MonoBehaviour
             }
             health -= x;
             Text txt = text.gameObject.GetComponent<Text>();
-            txt.text = (health / maxHealth) * 100 + "%";
+            txt.text = ((int)(health / maxHealth) * 100) + "%";
         }
     }
 
@@ -160,13 +160,13 @@ public class ShipHealth : MonoBehaviour
         {
             health = maxHealth;
             Text txt = text.gameObject.GetComponent<Text>();
-            txt.text = (health / maxHealth)*100 + "%";
+            txt.text = ((int)(health / maxHealth) * 100) + "%";
         }
         else
         {
             maxHealth += x;
             Text txt = text.gameObject.GetComponent<Text>();
-            txt.text = (health / maxHealth) * 100 + "%";
+            txt.text = ((int)(health / maxHealth) * 100) + "%";
         }
     }
 }
